Use DownloadHandler behind the scenes
